Streaming kit build
This is a sled to fit a NanoPi R5C and a tbs2603se encoder. Iterations 1+2 also held a battery. Iteration 2 tried to mount modems as well. Versions 3+4 add a fan. Version 4 also uses a quick board I soldered up to step up the 5v usb to 12v for the encoder, and also allow for 1 cable to power the fan, nanopi, and encoder.  If anyone does use this, test your fans, the variation in amp draw between the fans I bought was quite large.